EBC-47 - fix issues found by the checkstyle test. Two issues left unfixed as I'm not sure the rule is entirely inline with our current standards.
Summary of unfixed issue: When a line is broken on a binary operator, the Checkstyle test errors, indicating that the operator should be surrounded by a single space. When the operator begins the line, however, it is preceded by the tabs at the beginning of the line which will likely be more that a single character. I believe the test needs to be fixed as the line is correct in terms of style per the Zend style guide.