Eb2cPayment: PayPal API debugging
Log the request messages being sent out for PayPal.
sprintf with decimal format for any amounts in the request instead of just casting to string which may result in an empty string which would not be valid. This may be a bit naive in terms of i18n so may need to be explored in more depth.
Fix the getSubtotal calls, must be Subtotal not SubTotal
Fix the getShippingAmount calls, method is on the shipping address objects not the quote.
Shorten up some of the variable names. I know descriptive names are good and all but there's no need for them to be that specific, e.g. $paypalGetExpressCheckoutResponseMessage -> $responseMessage.