EBC-249 Remove gift card configuration options
Added tests for the isConfigSuppressed function.
Filled out the isConfigSuppressed function.
Changed isConfigSuppressed to allow group to be empty.
Switched to checking for the getName method.
Added test for the tabs override.
Added a test for the overridden tabs renderer.
Overrode checkSectionPermissions to not display suppressed tabs.
Switched the order of the isConfigSuppressed call.
Altered the test to disable the Tabs model's constructor.
Removed extraneous lines. The functionality was moved to the isConfigSuppressed function.
Fixed the class name.
Removed tests that were really for the suppression model.
Added tests for allowed/disallowed groups within the payment section.
Changed the altered tests to use strings as arguments to the FUT.
Added a docblock for the isConfigSuppressed function.
Removed more unecessary headers.
Changed isConfigSuppressed to accept only strings as arguments.
Updated the form test to make sure strings are passed to isConfigSuppressed.
Replaced the nested if statements with a boolean expression.
Fixed missing coverage in the tabs test case.
Rewrote the form override to not copy any magento code.
Updated the test.
Added a test to check the case where the group is suppressed.
Added a test to check the case where the node is not a child of the groups node.
Removed unnecessary lines and renamed tests.
Fixed a logic error and switched to using xpath for traversing.
Rewrote the suppression tests to be done with a dataprovider.
Removed unecessary fixtures.