EBC-316: converted the mysql4_paypal model into a proper resource model.
changed the models to inherit from the Db abstract resource models
instead of the mysql4 model.
rewrote the loadByQuoteId function to use the builtin ability to load a
model by another field.
renamed the model and move it to the resource subdirectory.
updated the config to reference the new resource model.