US5336 Configurable Handling of HTTP Responses
aliased the invokeCallback and getConfigData methods in the core helper.
added a helper with methods to be used as default callbacks for the api.
changed the other modules to be aware of, if not use, the changes to the api model.
changed the api model to log requests and responses at info level by default.
added locale file for the request failed message.

Conflicts:
	src/app/code/local/TrueAction/Eb2cCore/Helper/Data.php

Conflicts:
	src/app/code/local/TrueAction/Eb2cCore/Helper/Data.php
	src/app/code/local/TrueAction/Eb2cCore/Test/Helper/DataTest.php