DE829 41st Parameter Data in OrderCreate
Create frontend templates for Eb2cFraud to add 41st Parameter JS to OPC.
Add a new layout.xml file for Eb2cFraud. Add the frontend files to the
modman links.
Provide helper method to pull generated JS data from request POST data.
Move selection of 41st Parameter JS to run to the block used to output
the fraud data. Upon construction, the block will get the JS collector
to call and set the data as magic properties to be used in the template.
Delete the Jsc model that was previously doing all of this.